requirements.txt
requests>=2.23.0
logger
python-dotenv